Do not check if BluetoothAdapter is enabled or not while stopping the scans.
Otherwise we might end up having a ScanCallback that keeps getting called by OS despite the fact that Bluetooth is off. The way to reproduce it is to wait until we're in the middle of the scan, then turn Bluetooth off and attempt to stop the scan via IScanner.stop().